PR feedback and refactor
