+++ src/main/java/de/rwth/idsg/steve/repository/dto/ChargePoint.java
@@ -18,7 +18,7 @@
  */
 package de.rwth.idsg.steve.repository.dto;
 
-import de.rwth.idsg.steve.ocpp.ws.JsonObjectMapper;
+import de.rwth.idsg.steve.utils.JsonUtils;
 import de.rwth.idsg.steve.web.dto.ocpp.ConfigurationKeyEnum;
 import jooq.steve.db.tables.records.AddressRecord;
 import jooq.steve.db.tables.records.ChargeBoxRecord;
@@ -26,7 +26,6 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.joda.time.DateTime;
-import tools.jackson.databind.node.ObjectNode;
 
 /**
  *
@@ -50,24 +49,7 @@ public final class ChargePoint {
         private final AddressRecord address;
 
         public String getCpoName() {
-            return getConfigValueSafely(chargeBox, ConfigurationKeyEnum.CpoName.name());
-        }
-
-        private static String getConfigValueSafely(ChargeBoxRecord chargeBox, String key) {
-            var value = chargeBox.getOcppConfiguration();
-            if (value == null) {
-                return null;
-            }
-
-            var mapper = JsonObjectMapper.INSTANCE.getMapper();
-            var node = mapper.readTree(value.data());
-            if (!node.isObject()) {
-                return null;
-            }
-
-            var ocppConfiguration = (ObjectNode) node;
-            var configNode = ocppConfiguration.get(key);
-            return configNode == null ? null : configNode.asString();
+            return JsonUtils.getPropertyValueAsString(chargeBox.getOcppConfiguration(), ConfigurationKeyEnum.CpoName.name());
         }
     }
