+++ src/main/java/de/rwth/idsg/steve/repository/dto/ChargePointRegistration.java
@@ -19,6 +19,7 @@
 package de.rwth.idsg.steve.repository.dto;
 
 import de.rwth.idsg.steve.ocpp.OcppSecurityProfile;
+import de.rwth.idsg.steve.utils.JsonUtils;
 import de.rwth.idsg.steve.web.dto.ocpp.ConfigurationKeyEnum;
 import org.jetbrains.annotations.NotNull;
 import tools.jackson.databind.node.ObjectNode;
@@ -34,7 +35,6 @@ public record ChargePointRegistration(
 ) {
 
     public String cpoName() {
-        var node = ocppConfiguration.get(ConfigurationKeyEnum.CpoName.name());
-        return node == null ? null : node.asString();
+        return JsonUtils.getPropertyValueAsString(ocppConfiguration, ConfigurationKeyEnum.CpoName.name());
     }
 }
